[jnc_ct] fix: LlvmIrBuilder::createSwitch (intptr_t -> int64_t)

+++ src/jnc_ct/jnc_ct_ControlFlowMgr/jnc_ct_ControlFlowMgr_Eh.cpp
@@ -416,7 +416,7 @@ void ControlFlowMgr::finalizeFinallyScope(Scope* scope) {
 	}
 
 	char buffer1[256];
-	sl::Array<intptr_t> routeIdxArray(rc::BufKind_Stack, buffer1, sizeof(buffer1));
+	sl::Array<int64_t> routeIdxArray(rc::BufKind_Stack, buffer1, sizeof(buffer1));
 	routeIdxArray.setCount(count);
 
 	char buffer2[256];
